Move InfoBox component into PlayerControls directory
Not used except inside `PlayerControls` component.
